requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.0
bcrypt==4.1.2
